refactor(auth): reorganize auth modules and fix circular imports
- Move password hashing to app/services/auth/security.py
- Move core auth to app/auth/base.py
- Update imports across affected files
