+++ margin_app/app/auth/base.py
@@ -1,22 +1,18 @@
 """
-This module contains authentication related services.
+Base authentication module containing core authentication functionality.
 """
 
 from datetime import datetime, timedelta, timezone
-
 import jwt
 from jwt.exceptions import InvalidTokenError
-from passlib.context import CryptContext
 from starlette.requests import Request
+import requests
 
 from app.core.config import settings
 from app.crud.admin import admin_crud
 from app.models.admin import Admin
 from app.schemas.admin import AdminResponse
-from margin_app.auth.security import get_password_hash
-import requests
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+from app.services.auth.security import verify_password
 
 def get_expire_time(minutes: int) -> datetime:
     """
@@ -44,7 +40,6 @@ def create_access_token(email: str, expires_delta: timedelta | None = None):
     to_encode = {"sub": email, "exp": expire}
     return jwt.encode(to_encode, settings.secret_key, algorithm=settings.algorithm)
 
-
 async def get_current_user(token: str) -> Admin:
     """
     Retrieves the current user based on the provided JWT token.
@@ -60,7 +55,6 @@ async def get_current_user(token: str) -> Admin:
     - Exception("User not found"): if jwt correct but user doesn't exists
     - Exception("jwt expired"): if jwt expired
     """
-
     try:
         payload = jwt.decode(
             token, settings.secret_key, algorithms=[settings.algorithm]
@@ -75,26 +69,10 @@ async def get_current_user(token: str) -> Admin:
     except InvalidTokenError as e:
         raise Exception("jwt expired") from e
 
-
-def verify_password(plain_password, hashed_password) -> bool:
-    """
-    Verify that a plain text password matches a hashed password.
-
-    :param plain_password: str - The plain password to be verified.
-    :param hashed_password: str - The hashed password against which the plain
-    one will be verified.
-
-    :return: bool - True if the plain password matches the hashed password,
-    False if not.
-    """
-    return pwd_context.verify(plain_password, hashed_password)
-
-
 class GoogleAuth:
     """
     Google authentication service.
     """
-
     google_login_url = (
         f"https://accounts.google.com/o/oauth2/auth?"
         f"response_type=code&client_id={settings.google_client_id}"
@@ -161,9 +139,7 @@ class GoogleAuth:
             )
         return AdminResponse(id=user.id, name=user.name, email=user.email)
 
-
 google_auth = GoogleAuth()
-
 
 async def get_admin_user_from_state(req: Request) -> Admin | None:
     """
@@ -173,4 +149,4 @@ async def get_admin_user_from_state(req: Request) -> Admin | None:
 
     :return: Admin | None - The admin user if it exists, None otherwise.
     """
-    return getattr(req.state, "admin_user", None)
+    return getattr(req.state, "admin_user", None)
