+++ margin_app/app/services/auth/__init__.py
@@ -1,0 +1,3 @@
+"""
+Auth services package.
+"""
